Refactored: Chain of resp. Logic
