+++ src/main/java/de/uhd/ifi/se/decision/management/jira/consistency/checktriggers/StatusClosedTrigger.java
@@ -4,10 +4,9 @@ import com.atlassian.jira.event.issue.IssueEvent;
 import com.atlassian.jira.event.type.EventType;
 
 
-public class StatusClosedTrigger implements ConsistencyCheckEventTrigger, TriggerChain<ConsistencyCheckEventTrigger> {
+public class StatusClosedTrigger extends TriggerChain {
 
 	private IssueEvent issueEvent;
-	private TriggerChain nextChainLink;
 
 	public StatusClosedTrigger(IssueEvent issueEvent) {
 		this.issueEvent = issueEvent;
@@ -36,30 +35,7 @@ public class StatusClosedTrigger implements ConsistencyCheckEventTrigger, Trigge
 		return "closed";
 	}
 
-	@Override
-	public TriggerChain<ConsistencyCheckEventTrigger> setNextChain(TriggerChain<ConsistencyCheckEventTrigger> chain) {
-		this.nextChainLink = chain;
-		return chain;
-	}
 
-	@Override
-	public TriggerChain getNextChain() {
-		return this.nextChainLink;
-	}
-
-	@Override
-	public ConsistencyCheckEventTrigger getChainBase() {
-		return this;
-	}
-
-	@Override
-	public boolean calculate() {
-		boolean activated = isTriggered() && isActivated();
-		if (!activated && getNextChain() != null) {
-			activated = getNextChain().calculate();
-		}
-		return activated;
-	}
 
 
 }
