+++ src/main/java/de/uhd/ifi/se/decision/management/jira/consistency/checktriggers/WorkflowDoneTrigger.java
@@ -3,10 +3,9 @@ package de.uhd.ifi.se.decision.management.jira.consistency.checktriggers;
 
 import com.atlassian.jira.event.issue.IssueEvent;
 
-public class WorkflowDoneTrigger implements ConsistencyCheckEventTrigger, TriggerChain<ConsistencyCheckEventTrigger> {
+public class WorkflowDoneTrigger extends TriggerChain {
 
 	private IssueEvent issueEvent;
-	private TriggerChain nextChainLink;
 
 	public WorkflowDoneTrigger(IssueEvent issueEvent) {
 		this.issueEvent = issueEvent;
@@ -38,29 +37,5 @@ public class WorkflowDoneTrigger implements ConsistencyCheckEventTrigger, Trigge
 		return "done";
 	}
 
-	@Override
-	public ConsistencyCheckEventTrigger getChainBase() {
-		return this;
-	}
 
-
-	@Override
-	public TriggerChain<ConsistencyCheckEventTrigger> setNextChain(TriggerChain<ConsistencyCheckEventTrigger> chain) {
-		this.nextChainLink = chain;
-		return chain;
-	}
-
-	@Override
-	public TriggerChain getNextChain() {
-		return this.nextChainLink;
-	}
-
-	@Override
-	public boolean calculate() {
-		boolean activated = isTriggered() && isActivated();
-		if (!activated && getNextChain() != null) {
-			activated = getNextChain().calculate();
-		}
-		return activated;
-	}
 }
